Move actions related to tpm lib's permissions to provider
It could be called by other tests.
